Restructure and renaming files and composables
