--- app/src/main/java/hiof/gruppe15/treningsappen/ui/component/home/AnalyticsScreen.kt
@@ -1,23 +1,0 @@
-package hiof.gruppe15.treningsappen.ui.component.home
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
-import hiof.gruppe15.treningsappen.ui.component.navigation.AppScaffold
-
-@Composable
-fun AnalyticsScreen(navController: NavController) {
-    AppScaffold(navController = navController, title = "Analytics") {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(it)
-        ) {
-            Text("Analytics")
-        }
-    }
-}
