+++ app/src/main/java/hiof/gruppe15/treningsappen/ui/component/navigation/NavGraph.kt
@@ -4,35 +4,33 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import hiof.gruppe15.treningsappen.ui.component.auth.ForgotPasswordScreen
-import hiof.gruppe15.treningsappen.ui.component.auth.LoginScreen
-import hiof.gruppe15.treningsappen.ui.component.auth.RegisterScreen
-import hiof.gruppe15.treningsappen.ui.component.home.AnalyticsScreen
+import hiof.gruppe15.treningsappen.ui.component.login.ForgotPasswordScreen
+import hiof.gruppe15.treningsappen.ui.component.login.LoginScreen
+import hiof.gruppe15.treningsappen.ui.component.login.RegisterScreen
+import hiof.gruppe15.treningsappen.ui.component.analytics.AnalyticsScreen
 import hiof.gruppe15.treningsappen.ui.component.home.HomeScreen
 import hiof.gruppe15.treningsappen.ui.component.home.SaveRoutineScreen
 import hiof.gruppe15.treningsappen.ui.component.profile.ProfileScreen
-import hiof.gruppe15.treningsappen.ui.component.settings.SettingsScreen
-import hiof.gruppe15.treningsappen.ui.component.workout.RoutineScreen
-import hiof.gruppe15.treningsappen.ui.component.workout.WorkoutSessionScreen
+import hiof.gruppe15.treningsappen.ui.component.routines.RoutineScreen
+import hiof.gruppe15.treningsappen.ui.component.routines.WorkoutSessionScreen
 
 sealed class Screen(val route: String) {
     object Login : Screen("login")
     object Register : Screen("register")
     object ForgotPassword : Screen("forgotPassword")
     object Home : Screen("home")
-    object SaveTrainingRoutine : Screen("saveTrainingRoutine")
-    object Profile : Screen("profile")
-    object Settings : Screen(route = "settings")
-    object WorkoutPlan : Screen(route = "workoutPlan")
+    object Routines : Screen(route = "workoutPlan")
     object WorkoutSession : Screen(route = "workoutSession")
+    object CreateNewRoutine : Screen("createNewRoutine")
     object Analytics : Screen(route = "analytics")
+    object Profile : Screen("profile")
 }
 
 sealed class ScreenCategory(val routes: List<String>) {
-    object Home : ScreenCategory(listOf("home", "saveTrainingRoutine"))
-    object Workout : ScreenCategory(listOf("workoutPlan", "workoutSession"))
+    object Home : ScreenCategory(listOf("home"))
+    object Workout : ScreenCategory(listOf("workoutPlan", "workoutSession", "createNewRoutine"))
     object Analytics : ScreenCategory(listOf("analytics"))
-    object Profile : ScreenCategory(listOf("profile", "settings"))
+    object Profile : ScreenCategory(listOf("profile"))
 }
 
 @Composable
@@ -42,15 +40,18 @@ fun NavGraph(
     NavHost(
         navController = navController, startDestination = startDestination
     ) {
-        composable(Screen.Register.route) { RegisterScreen(navController) }
         composable(Screen.Login.route) { LoginScreen(navController) }
-        composable(Screen.Home.route) { HomeScreen(navController) }
-        composable(Screen.Profile.route) { ProfileScreen(navController) }
-        composable(Screen.Settings.route) { SettingsScreen(navController) }
-        composable(Screen.WorkoutPlan.route) { RoutineScreen(navController) }
-        composable(Screen.WorkoutSession.route) { WorkoutSessionScreen(navController) }
-        composable(Screen.Analytics.route) { AnalyticsScreen(navController) }
+        composable(Screen.Register.route) { RegisterScreen(navController) }
         composable(Screen.ForgotPassword.route) { ForgotPasswordScreen(navController) }
-        composable(Screen.SaveTrainingRoutine.route) { SaveRoutineScreen(navController) }
+
+        composable(Screen.Home.route) { HomeScreen(navController) }
+
+        composable(Screen.Routines.route) { RoutineScreen(navController) }
+        composable(Screen.WorkoutSession.route) { WorkoutSessionScreen(navController) }
+        composable(Screen.CreateNewRoutine.route) { SaveRoutineScreen(navController) }
+
+        composable(Screen.Analytics.route) { AnalyticsScreen(navController) }
+
+        composable(Screen.Profile.route) { ProfileScreen(navController) }
     }
 }
